rewrote updating list of assets when editing
