+++ src/actions/types.js
@@ -84,6 +84,11 @@ export const USER_REMOVED_ASSET_FROM_LIST_SUCCESS =
   '[user] REMOVED_ASSET_FROM_LIST'
 export const USER_REMOVED_ASSET_FROM_LIST_FAILED =
   '[user] REMOVED_ASSET_FROM_LIST'
+export const USER_EDIT_ASSETS_IN_LIST = '[assets] EDIT_ASSETS_IN_LIST'
+export const USER_EDIT_ASSETS_IN_LIST_SUCCESS =
+  '[assets] EDIT_ASSETS_IN_LIST_SUCCESS'
+export const USER_EDIT_ASSETS_IN_LIST_FAILED =
+  '[assets] EDIT_ASSETS_IN_LIST_FAILED'
 export const ASSETS_FETCH = '[assets] ASSETS_FETCH'
 export const ASSETS_FETCH_SUCCESS = '[assets] ASSETS_FETCH_SUCCESS'
 export const ASSETS_FETCH_FAILED = '[assets] ASSETS_FETCH_FAILED'
